@@ -65,7 +65,7 @@
     "@vue/compiler-sfc": "^3.1.0",
     "cross-env": "^5.2.0",
     "minimist": "^1.2.0",
-    "node-sass": "^7.0.0",
+    "node-sass": "^7.0.1",
     "rollup": "^2.75.6",
     "rollup-plugin-livereload": "^1.3.0",
     "rollup-plugin-postcss": "^3.1.2",
@@ -77,7 +77,7 @@
     "rollup-plugin-serve": "^1.1.0",
     "rollup-plugin-vue": "^6.0.0",
     "sass-loader": "^13.0.0",
-    "vue": "^3.1.0",
+    "vue": "3.2.37",
     "vue-knob-control": "^1.7.0"
   },
   "dependencies": {
